added immediate arithmetic instruction disassembly and rudimentary roms/instruction memo ry. created a rom generation script to test disassembly

+++ riscv-simulator/src/rom.rs
@@ -40,7 +40,7 @@ impl Rom {
     }
 
     pub fn get_instruction(&self, address: usize) -> u32 {
-        return self.data[address];
+        return self.data[address as usize];
     }
 
 }
